// frontend-2/src/lib/types/property.ts
export type Property = {
  id?: number;
  user_id: number;
  title: string;
  description: string;
  price: string;
  bedrooms: number;
  bathrooms: number;
  square_feet: number;
  lot_size: string;
  year_built: number;
  status: string;
  listed_date: string;
  hoa_fees: string;
  location_id: number;
  city: string;
  country: string;
};